start_service_starlite.py
# Starts a skir service on http://localhost:8787/?myapi
#
# Run with:
#   litestar --app start_service_starlite:app run --host localhost --port 8787
#
# Run call_service.py to call this service from another process.

import urllib.parse

import skir
from litestar import Litestar, Request, Response, get, route

from skirout import service_skir, user_skir


class ServiceImpl:
    def __init__(self):
        self._id_to_user = {}

    async def get_user(
        self, request: service_skir.GetUserRequest
    ) -> service_skir.GetUserResponse:
        user_id = request.user_id
        user = self._id_to_user.get(user_id)
        return service_skir.GetUserResponse(user=user)

    async def add_user(
        self,
        request: service_skir.AddUserRequest,
        req_headers: dict[str, str],
    ) -> service_skir.AddUserResponse:
        user = request.user
        if user.user_id == 0:
            raise ValueError("invalid user id")
        print(f"Adding user: {user}")
        self._id_to_user[user.user_id] = user
        return service_skir.AddUserResponse()

    _id_to_user: dict[int, user_skir.User]


service_impl = ServiceImpl()

skir_service = skir.ServiceAsync[dict[str, str]]()
skir_service.add_method(service_skir.AddUser, service_impl.add_user)
skir_service.add_method(service_skir.GetUser, service_impl.get_user)


@get("/")
async def hello_world() -> str:
    return "Hello, World!"


@route("/myapi", http_method=["GET", "POST"])
async def myapi(request: Request) -> Response:
    if request.method == "POST":
        req_body = (await request.body()).decode("utf-8")
    else:
        query_string = request.scope.get("query_string", b"").decode("utf-8")
        req_body = urllib.parse.unquote(query_string)
    req_headers = dict(request.headers)
    raw_response = await skir_service.handle_request(req_body, req_headers)
    return Response(
        content=raw_response.data,
        status_code=raw_response.status_code,
        media_type=raw_response.content_type,
    )


app = Litestar(route_handlers=[hello_world, myapi])
